Allow to cancel before generating image
